+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/validation/DefaultLibraryAliasValidator.java
@@ -2,10 +2,10 @@ package uk.ac.bbsrc.tgac.miso.core.service.naming.validation;
 
 public class DefaultLibraryAliasValidator extends RegexValidator {
 
-  private static final String regex = "([A-z0-9]+)_L([A-z0-9]+)-([A-Z0-9]+)_(.*)";
+  public static final String REGEX = "([A-z0-9]+)_L([A-z0-9]+)-([A-Z0-9]+)_(.*)";
 
   public DefaultLibraryAliasValidator() {
-    super(regex, false, false);
+    super(REGEX, false, false);
   }
 
   @Override