+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/BulkLibraryAliquotIT.java
@@ -25,14 +25,15 @@ import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.element.HandsOnTable;
 
 public class BulkLibraryAliquotIT extends AbstractIT {
 
-  private static final Set<String> columns = Sets.newHashSet(LibraryAliquotColumns.NAME, LibraryAliquotColumns.ALIAS,
-      LibraryAliquotColumns.PARENT_ALIAS, LibraryAliquotColumns.ID_BARCODE, LibraryAliquotColumns.BOX_SEARCH,
-      LibraryAliquotColumns.BOX_ALIAS, LibraryAliquotColumns.BOX_POSITION, LibraryAliquotColumns.DISCARDED,
-      LibraryAliquotColumns.EFFECTIVE_GROUP_ID, LibraryAliquotColumns.GROUP_ID, LibraryAliquotColumns.GROUP_DESCRIPTION,
-      LibraryAliquotColumns.DESIGN_CODE, LibraryAliquotColumns.SIZE, LibraryAliquotColumns.CONCENTRATION,
-      LibraryAliquotColumns.CONCENTRATION_UNITS, LibraryAliquotColumns.VOLUME, LibraryAliquotColumns.VOLUME_UNITS,
-      LibraryAliquotColumns.NG_USED, LibraryAliquotColumns.VOLUME_USED, LibraryAliquotColumns.CREATION_DATE,
-      LibraryAliquotColumns.TARGETED_SEQUENCING);
+  private static final Set<String> commonColumns = Sets.newHashSet(LibraryAliquotColumns.NAME, LibraryAliquotColumns.ALIAS,
+      LibraryAliquotColumns.ID_BARCODE, LibraryAliquotColumns.BOX_SEARCH, LibraryAliquotColumns.BOX_ALIAS,
+      LibraryAliquotColumns.BOX_POSITION, LibraryAliquotColumns.DISCARDED, LibraryAliquotColumns.EFFECTIVE_GROUP_ID,
+      LibraryAliquotColumns.GROUP_ID, LibraryAliquotColumns.GROUP_DESCRIPTION, LibraryAliquotColumns.DESIGN_CODE,
+      LibraryAliquotColumns.SIZE, LibraryAliquotColumns.CONCENTRATION, LibraryAliquotColumns.CONCENTRATION_UNITS,
+      LibraryAliquotColumns.VOLUME, LibraryAliquotColumns.VOLUME_UNITS, LibraryAliquotColumns.NG_USED, LibraryAliquotColumns.VOLUME_USED,
+      LibraryAliquotColumns.CREATION_DATE, LibraryAliquotColumns.TARGETED_SEQUENCING);
+
+  private static final Set<String> propagateColumns = Sets.newHashSet(LibraryAliquotColumns.PARENT_ALIAS);
 
   private static final String NO_TAR_SEQ = "(None)";
 
@@ -47,8 +48,8 @@ public class BulkLibraryAliquotIT extends AbstractIT {
     BulkLibraryAliquotPage page = BulkLibraryAliquotPage.getForEdit(getDriver(), getBaseUrl(), Sets.newHashSet(304L, 305L));
     HandsOnTable table = page.getTable();
     List<String> headings = table.getColumnHeadings();
-    assertEquals(columns.size(), headings.size());
-    for (String col : columns) {
+    assertEquals(commonColumns.size(), headings.size());
+    for (String col : commonColumns) {
       assertTrue("Check for column: '" + col + "'", headings.contains(col));
     }
     assertEquals(2, table.getRowCount());
@@ -61,8 +62,11 @@ public class BulkLibraryAliquotIT extends AbstractIT {
         Sets.newHashSet(304L));
     HandsOnTable table = page.getTable();
     List<String> headings = table.getColumnHeadings();
-    assertEquals(columns.size(), headings.size());
-    for (String col : columns) {
+    assertEquals(commonColumns.size() + propagateColumns.size(), headings.size());
+    for (String col : commonColumns) {
+      assertTrue("Check for column: '" + col + "'", headings.contains(col));
+    }
+    for (String col : propagateColumns) {
       assertTrue("Check for column: '" + col + "'", headings.contains(col));
     }
     assertEquals(1, table.getRowCount());
@@ -77,8 +81,11 @@ public class BulkLibraryAliquotIT extends AbstractIT {
         Sets.newHashSet(304L));
     HandsOnTable table = page.getTable();
     List<String> headings = table.getColumnHeadings();
-    assertEquals(columns.size(), headings.size());
-    for (String col : columns) {
+    assertEquals(commonColumns.size() + propagateColumns.size(), headings.size());
+    for (String col : commonColumns) {
+      assertTrue("Check for column: '" + col + "'", headings.contains(col));
+    }
+    for (String col : propagateColumns) {
       assertTrue("Check for column: '" + col + "'", headings.contains(col));
     }
     assertEquals(1, table.getRowCount());